added notebook and fixed directory
