--- code/utils/description_pp/sample_description.json
@@ -1,1 +1,0 @@
-{"book": [], "cup": [], "people": []}
